Move all metadata decoding to the same file and export only one method and the relevant types.
