+++ ts/src/decoding/stringDecoder.ts
@@ -1,4 +1,4 @@
-import { decodeStreamMetadataExtended } from "../metadata/tile/streamMetadataDecoder";
+import { decodeStreamMetadata } from "../metadata/tile/streamMetadataDecoder";
 import { StringFlatVector } from "../vector/flat/stringFlatVector";
 import { StringDictionaryVector } from "../vector/dictionary/stringDictionaryVector";
 import type IntWrapper from "./intWrapper";
@@ -11,7 +11,6 @@ import { decodeIntStream, decodeLengthStreamToOffsetBuffer, decodeNullableIntStr
 import { type Column, ScalarType } from "../metadata/tileset/tilesetMetadata";
 import { decodeVarintInt32 } from "./integerDecodingUtils";
 import { decodeBooleanRle, skipColumn } from "./decodingUtils";
-import { RleEncodedStreamMetadata } from "../metadata/tile/rleEncodedStreamMetadata";
 import { StringFsstDictionaryVector } from "../vector/fsst-dictionary/stringFsstDictionaryVector";
 
 const ROOT_COLUMN_NAME = "default";
@@ -34,7 +33,7 @@ export function decodeString(
     let plainDataStream: Uint8Array = null;
 
     for (let i = 0; i < numStreams; i++) {
-        const streamMetadata = decodeStreamMetadataExtended(data, offset);
+        const streamMetadata = decodeStreamMetadata(data, offset);
         if (streamMetadata.byteLength === 0) {
             continue;
         }
@@ -195,7 +194,7 @@ export function decodeSharedDictionary(
 
     let dictionaryStreamDecoded = false;
     while (!dictionaryStreamDecoded) {
-        const streamMetadata = decodeStreamMetadataExtended(data, offset);
+        const streamMetadata = decodeStreamMetadata(data, offset);
         switch (streamMetadata.physicalStreamType) {
             case PhysicalStreamType.LENGTH:
                 if (LengthType.DICTIONARY === streamMetadata.logicalStreamType.lengthType) {
@@ -248,9 +247,9 @@ export function decodeSharedDictionary(
             throw new Error("Currently only optional string fields are implemented for a struct.");
         }
 
-        const presentStreamMetadata = decodeStreamMetadataExtended(data, offset);
+        const presentStreamMetadata = decodeStreamMetadata(data, offset);
         const presentStream = decodeBooleanRle(data, presentStreamMetadata.numValues, offset);
-        const offsetStreamMetadata = decodeStreamMetadataExtended(data, offset);
+        const offsetStreamMetadata = decodeStreamMetadata(data, offset);
         const offsetCount = offsetStreamMetadata.decompressedCount;
         const isNullable = offsetCount !== numFeatures;
         const offsetStream = isNullable
